requirements.txt
requests
pandas
mysql-connector-python
oauth2client